requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
ansicon==1.89.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==19.3.0
Babel==2.11.0
backcall==0.2.0
bleach==4.1.0
blessed==1.19.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
contextvars==2.4
cycler==0.11.0
dataclasses
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dictdiffer==0.9.0
diskcache==5.4.0
docopt==0.6.2
dtlpy==1.71.38
entrypoints==0.4
filelock==3.4.1
frozenlist==1.2.0
fuzzyfinder==2.1.0
idna==3.4
idna-ssl==1.1.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
inquirer==2.8.0
ipdb==0.13.11
ipykernel==5.5.6
ipython
ipython-genutils
ipywidgets
jedi==0.17.2
Jinja2==3.0.3
jinxed==1.2.0
json5==0.9.11
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numba==0.53.1
numpy==1.21.6
open3d
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
pycparser==2.21
pydantic==1.9.2
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.7.1
PyYAML
pyzmq
readchar
redis==4.3.5
requests==2.27.1
requests-toolbelt==0.9.1
scipy==1.5.4
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
terminado==0.12.1
testpath==0.6.0
tomli==1.2.3
tomlkit==0.11.6
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.14
validators==0.18.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
webvtt-py==0.4.3
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.6.0
dtlpy